[BUGFIX] Do not crash linkvalidator report with missing TCA field label
The mask extension[1] automatically generates column definitions in TCA,
and they do not have a 'label' key in the TCA columns definition.

In TYPO3 v11.5.37, the link validator report page crashes with the
following error:
> (1/1) #1476107295 TYPO3\CMS\Core\Error\Exception
> PHP Warning: Undefined array key "label"
> in [...]/linkvalidator/Classes/Report/LinkValidatorReport.php line 559

In TYPO3 v12 the problem does not occur anymore.

This patch simply checks if the label key exists before it is used.

[1] https://github.com/Gernott/mask

Resolves: #104032
Releases: 11.5
Change-Id: Icca6f102883b09b3fec182158e625b959988db33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84592
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>