[TASK] Provide request object to EXT:felogin events
In order to enable more possibilities in the events,
the PSR-14 events in EXT:felogin are extended with
the PSR-7 request object, otherwise the events are useless
as $GLOBALS['TYPO3_REQUEST'] must be used.

Currently, it is f.e. not possible to modify the view
ModifyLoginFormViewEvent based e.g. on a site configuration.

Resolves: #104100
Releases: main, 12.4
Change-Id: I9daa6481d71fd1cae98d0a7ab01fe1adc7602748
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84696
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>