[BUGFIX] Convert fixed font sizes to relative units
In continuation of #104044 fixed font size definition on
certain components are migrated to use relative font units.

Note that CSS that is emitted to the TYPO3 frontend uses
`em` instead of `rem` in order to avoid putting a demand on
the font size that is defines on the frontend specific HTML
tag styling. Instead the current font size of the context
where the respective element is rendered is taken into account
for relative sizing.

Resolves: #104096
Related: #104044
Releases: main
Change-Id: Ia29fe95fb96709df70d2df439b4d89aaf0c1e355
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84697
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>