[TASK] Render generator fields as hidden fields in columnsOnly mode
When using the "columnsOnly" mode to render just
a subset of available fields of a record, while
the subset includes a field of TCA type "slug",
FormEngine needs to also render configured generator
fields. Otherwise the slug fields won't work as
expected, e.g. when recalculating.

However, since it might be confusing for an editor
to get fields rendered, which were not selected,
are those fields now rendered as hidden fields.

This is done by adding those fields to a hidden
palette while omitting duplicates and performing
sanitization.

Resolves: #104115
Releases: main, 12.4
Change-Id: Ia31571bae9913028e81df910462d76dc4314644c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84721
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: core-ci <typo3@b13.com>