[FEATURE] Add setting for backend lock file location, drop typo3conf
As a follow-up to #104074, the directory "typo3conf" is no longer
required to store the `LOCK_BACKEND` file.

A new configuration option TYPO3_CONF_VARS[BE][lockBackendFile]
is added alongside a distinct `BackendLocker` service class to
indicate where the backend lock and unlock commands store
their lockfile. By default this uses `var/lock/` for composer mode,
and `config/` for legacy installations as a fallback for when
the setting is empty (default).

Resolves: #104126
Related: #104074
Related: #104119
Releases: main
Change-Id: I1fdbd48434c2c6cfd0a6617a1be71ac21195a8e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84729
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>