[BUGFIX] Show last page of live search with different result-groups
If the last page of the live search consists of entries from different
search-providers like Pages and Records it renders correctly now.

Before the counting of total-count of items was wrong. Thereby
`array_fill()` was called with a negative value for the parameter
`count` which resulted in a `ValueError`.

Resolves: #104149
Releases: main, 13.4, 12.4
Change-Id: I72a95dffe06303385cac0f063ae0e1e17c8c58fb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86899
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>