[BUGFIX] Fix image manipulation GUI only operating on 1st crop variant
When multiple crop variants are defined (for example in EXT:styleguide
"Form engine elements - imageManipulation" for "crop_5" with
"mobile" and "desktop), the currently selected aspect ratio
and the size of the cropping box was only computed based on the
first crop variant.

This patch uses a wrapping container for each variant and adjusts
the JavaScript to operate on that context.

The information for the current cropping box size is now also
based on this context.

Resolves: #103879
Releases: main, 12.4
Change-Id: I38e235ddb5ed5f7463a5f0931608944326097f9f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84781
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>