[TASK] Allow additional composer options for `-s composerInstall`
This change modifies the `Build/Script/runTests.sh`
command dispatcher to allow passing composer options
to the `composer install` subcommand, for example

  Build/Scripts/runTests.sh -s composerInstall \
    -- --prefer-source

Note that this is already possible using the generic
`-s composer` dispatcher but would resolve some user
confusion.

This is useful to install dependencies as git
checkouts, for example when working on packages
like `typo3/testing-framework` changes. Albeit
the need to remove the vendor folder first is
needed in case of the example.

Resolves: #104186
Releases: main, 12.4, 11.5
Change-Id: Id07c90f3b6902ad8866249563eceeb2dafbc919b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84845
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>