[BUGFIX] Silence regexp pattern errors in EXT:redirects
This change adds a `@` silence operator to the usage of
`preg_match()` in the `RedirectService`. This prevents errors
in case of manually-entered malformed regexp patterns and follows
existing similar usages of the service.

Additionally, a todo comment is added to evaluate future early
pattern evaluation in record creation/editing so that
a proper response to invalid patterns can be emitted.

Resolves: #102176
Releases: main, 12.4, 11.5
Change-Id: I88687c128986ae6fa1e2b7a3a4bf414ee8f952f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84872
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>