[BUGFIX] Allow using parentheses in TypoScript operator functions
The TypoScript Parser was not able to parse this:

foo := appendString( AND `uid` NOT IN (48,10,50))

It recogized the first closing parentheses ")" as
function end. The patch fixes this by counting the
number of parentheses to avoid early stop. This
approach is limited, but it establishes backwards
compatibility for most use cases.

A better solution is to have "inner" parentheses
quoted to cope with cases where the number of opening
and closing parentheses don't match, but that's a syntax
addition that needs to be done with a separate patch.

Limits of current approach: In case of extra closing
parentheses, any string is ignored past the first valid
match:

foo = Do
foo := appendString( something(withExtra())))))))))))
// Result: "Do something(withExtra())"

In case of extra opening parantheses and no matching
closing parentheses, the result will be an
empty string/stream:

foo = Do
foo := appendString( something(((((((()
// Result: "Do"

Note multiline assignments remain invalid:

foo = Do
foo := appendString(
something
)

Resolves: #101636
Related: #102742
Related: #97816
Related: #103671
Releases: main, 13.4, 12.4
Change-Id: Ia7352584c1139af60c3dd127cbbe701c9efc0c75
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86985
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>