[BUGFIX] Allow empty field to override default in DataHandler
Given a TCA type "select" with various options. One option using the
value of "". If a TCA config key "default" is set, which is different
from an empty string, it has been not possible to select the option
with an empty string as value.

As a solution, it is checked if the value handed to DataHandler is part
of the available option items and their values before using the default
config value as fallback.

Resolves: #95428
Releases: main, 13.4
Change-Id: Ic17bd39c08fd4f49c3bb8f15196f457ad1657c64
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88266
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>