[TASK] Avoid implicitly nullable class method parameter in `EXT:indexed_search`
With PHP 8.4 marking method parameter implicitly nullable
is deprecated and will emit a `E_DEPRECATED` warning. One
recommended way to resolve this, is making it explicitly
nullable using the `?` nullable operator or adding a null
tyype to an union type definition. [1]

This prepares the way towards PHP 8.4 compatibility.

[1] https://php.watch/versions/8.4/implicitly-marking-parameter-type-nullable-deprecated

Resolves: #104236
Releases: main, 12.4, 11.5
Change-Id: Iba5e99b526ca37d7bb36044cd6ee1027b6c5245a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84951
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>