[TASK] Make PagePermissionAssembler stateless
PagePermissionAssembler is a DataHandler related
helper that calculates the permission fields of
page records.

The patch removes pollution of __construct() by
manual arguments: $GLOBALS['TYPO3_CONF_VARS'] is
runtime static and can just be used when needed.
This avoids a class property, making the class
stateless.

Minor further clean ups, an improved comment, and
an @todo that we should consider changing the API.

Resolves: #104245
Releases: main
Change-Id: I85d125612af02b54627a4c9efd525d4349ce868c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84985
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>