[TASK] Use TcaSchema in ReferenceIndex
Avoid $GLOBALS['TCA'] access in ReferenceIndex
by using TcaSchema. This is a patch to learn
about fiddly areas and pitfalls when using
TcaSchema in scenarios with lots of different
field related details and relations: The
general API feels good, but it's sensitive
to errors with such transition patches.

Resolves: #104366
Related: #104002
Releases: main
Change-Id: I3cce9cc9681348454ddb87fc5ebd90139dbc4776
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85008
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>