[TASK] Avoid implicitly nullable class method parameter in `EXT:adminpanel`
With PHP 8.4 marking method parameter implicitly nullable
is deprecated and will emit a `E_DEPRECATED` warning. One
recommended way to resolve this, is making it explicitly
nullable using the `?` nullable operator or adding a null
type to an union type definition. [1]

This prepares the way towards PHP 8.4 compatibility.

[1] https://php.watch/versions/8.4/implicitly-marking-parameter-type-nullable-deprecated

Resolves: #104266
Releases: main, 12.4, 11.5
Change-Id: I731d63dbbb13df6b3e0b6f40641ac0a98e92519f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85024
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>