[TASK] Fix CGL issues
Apply the latest rules of php-cs-fixer for
both fixer (code and header-comment) also
in the `Build/Scripts` folder.

The Build folder is now added to the usual
php-cs-fixer execution for code and header
fixer scripts to avoid code style drifitng
in this folder for the future.

Used command(s):

  Build/Scripts/runTests.sh -s cgl ; \
    Build/Scripts/runTests.sh -s cglHeader

Resolves: #104274
Releases: main, 12.4, 11.5
Change-Id: I7977e2254b9e271c45b840c6696cd7600ab0b1aa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85035
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>