[TASK] Use TcaPreparation in FlexFormTools
With introduction of new types "category", "file",
"email" and "link", class TcaPreparation has been
established to auto-add TCA details that should
be shadowed-away for devs, and are added internally.

This created quite a mess within FlexFormTools since
TcaPrepration needs to be applied there as well, but
with slightly different "category" and "file" details.

Also, "email" and "link" received some hacks in
runtime code to dynamically add "softref" - a thing
TcaPreparation should handle as well, but didn't,
also due to the hacks in FlexFormTools.

This is an issue in combination with new TcaSchema:
TcaSchemaFactory relies on TCA being "final" and
not changed at runtime.

The patch clean this up:

* Free FlexFormTools from own implementations of
  "category" and "file" fields, call TcaPreparation.
* Align TcaPreparation with the changes for flex
  fields by having a toggle for this and act upon.
* Move "email" and "link" "softref" runtime hacks
  into TcaPreparation.
* Protect a helper method in TcaPreparation
  previously used by FlexFormTools.
* Refactor a double nested closure in TcaPreparation
  towards methods.
* Refactor some tests and add some easy ones for
  "sofref" handling in TcaPreparation.

Resolves: #104293
Related: #104002
Related: #103316
Related: #97013
Related: #97159
Related: #94622
Related: #104287
Releases: main
Change-Id: If0cf7f99aad2acdedd785186c199252fb6388faf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85037
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>