[TASK] Avoid implicitly nullable class method parameter in `EXT:backend`
With PHP 8.4 marking method parameter implicitly nullable
is deprecated and will emit a `E_DEPRECATED` warning. One
recommended way to resolve this, is making it explicitly
nullable using the `?` nullable operator or adding a null
type to an union type definition. [1]

This prepares the way towards PHP 8.4 compatibility.

[1] https://php.watch/versions/8.4/implicitly-marking-parameter-type-nullable-deprecated

Resolves: #104278
Releases: main, 12.4, 11.5
Change-Id: Ia3a4bc344b369b8697c88ed14ac06354dd988dce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85041
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>