[TASK] Move TCA control sanitization to TcaPreparation
The functionality to ensure that system internal
columns that are required for data integrity
are available in case they have been defined
in the TCA control section, was previously
done in TcaMigration.

However, since TcaMigration should only
contain migration functionality, which
on the one hand logs deprecation warnings
and on the other hand get's removed again
after some versions, the corresponding
functionality is now moved to TcaPreparation.

Resolves: #104294
Releases: main
Change-Id: I5e072cef866b36440f9409b1c68e114892ebcd51
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85084
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>