[TASK] Improve UX and functionality of "Linkvalidator" report table
The report table contains broken links that exist in the system
depending on the configuration. The information about these broken
links can be varied and the previous view already provided essential
information. With this patch the view has been expanded and the UX has
been improved.

Information about the link text, (location) path in the page tree and
the record field that contains the broken link have been moved to a
modal so that the table only contains the relevant information and
further information can be viewed in the modal.

The new modal is divided into an area with information about the link
properties and an area with extensive information about the record
itself.

Furthermore a column has been added to the table that shows the
language of the record that contains the broken link.

Previously, the record icon was simply the standard icon of the table,
now the actual record type is taken into account as the icon.

A sorting function has also been added to the table with which the
columns can be sorted alphabetically in ascending/descending order.

Resolves: #104323
Releases: main
Change-Id: Id7279080c8f851c1b8cb7d45f4e0116fa6a1fc49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85088
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>