[BUGFIX] Add missing sortable folder timestamps to filelist view
With the introduction of the new filelist functionality,
the list view of folders no longer showed their last modification
timestamps. Without this information, the list view could no
longer sort the whole item list by date.

With this patch, the information is now fetched not only for
files, but also for folders.

Folders and files are sorted on their own, so no file will be
sorted in between a list of folders (and vice versa).

(Note that for filesystem folders there is only one timestamp,
so "last modified" and "created at" will use the same timestamps.
Remote folders with their distinct drivers may provide distinguishable
information.)

Resolves: #100520
Releases: main, 12.4
Change-Id: I4e9ceb597946cc7532b76ec2ee4bec3d52d17fdc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85193
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>