[TASK] Allow PHP 8.4 in runTests.sh
This change adds PHP8.4 as allowed version to the
`Build/Scripts/runTests.sh` script as preparation
to validate and check the Core and 3rd party libs
with PHP8.4 locally.

Note that unit tests already revealing deprecation
and errors for 3rd party extension which can now
be reported and for checked upstream updates.

Resolves: #104337
Releases: main, 12.4, 11.5
Change-Id: I76d3e37b393c347173fcdbe874d6a0737fb76d33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85180
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>