[BUGFIX] Prevent TypeError int values of backend user settings
PHP will cast array keys to int if they are valid integers.
The code applies htmlspecialchars() on the keys.
The function expects a string as first argument, which wouldn't be the
case if the actual value would be an int.

Therefore the key is casted to a string within the function call to
prevent TypeError.

Resolves: #104341
Releases: main, 12.4
Change-Id: Ic9aef0fd37f2e883a2a8fa626cfda2e813e3b289
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85190
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>