[BUGFIX] Provide luxon-compatible format token for aria labels in flatpickr
The flatpickr integration in TYPO3 is configured to use luxon for date
time formatting operations, which consumes different format tokens. This
leads to the situation that aria labels generated by flatpickr are now
broken.

This commit changes the option `ariaDateFormat` to luxon's `DDDD` token,
which renders a "localized date with full month and weekday", according
to [1].

[1] https://moment.github.io/luxon/#/formatting?id=table-of-tokens

Resolves: #104342
Releases: main, 12.4
Change-Id: I7902d39209ad59e0c3196db7f906b93533d7b7cf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85191
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>