[BUGFIX] Set `hidden` on toggling alert visibility state
Toggling the visibility of the new alert element,
introduced with #103699, does actually lead to
changing the `opacity` property. Therefore, the
corresponding space of the alert element will
stay, leaving empty spaces in the backend.

This is now resolved by additionally toggling the
`hidden` class, which works with the `display`
property.

Note: The `show` class has to stay due to the
underlying functionality from the used bootstrap
component.

Resolves: #104346
Related: #103699
Releases: main
Change-Id: Id85da0972be02b43ef78f79fd1118d74e8ff1662
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85213
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Marcin Sągol <marcin@soee.pl>
Reviewed-by: Marcin Sągol <marcin@soee.pl>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>