[BUGFIX] Properly handle boolean attribute in language packs
When using boolean attributes there are two things to
consider:

1. The default value must be FALSE
2. Presence of the attribute always means TRUE

It's therefore not possible to have a boolean
attribute, which defaults to TURE and it's
also not possible to set an attribute to
FALSE via its value (e.g. "0" or "false).

See for reference: https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#boolean-attributes

Therefore, to properly handle the "is writeable"
state in the language packs module, the
default is set to FALSE an the attribute
only gets added (which defines the TRUE state)
in case corresponding data attribute is
set in the template.

Resolves: #104353
Releases: main, 12.4
Change-Id: I82f440ae4fbeba405c35bbfa859804e2bc0660c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85217
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>