[TASK] Update avatar component page in "Styleguide" backend module
Naturally, ViewHelpers convert their contents immediately, which means
that in the examples of the "Styleguide" backend module, developers
only see the converted code and not the ViewHelper actually used. To
avoid this, code examples must use "non-intact" ViewHelpers. With this
patch, the CodeViewHelper is extended to decode transmitted entities.
As an example, the code examples on the "Avatar" component page have
been adapted accordingly.
In addition, the standard height in the CodeViewHelper's code mirror
configuration has been reduced.

Furthermore, new styling has been introduced for example uses and code
examples. This has also been adapted as an example on the "Avatar"
component page.

Finally, the texts on the "Avatar" component page have been updated.

Further adjustments to the other example pages in the "Styleguide"
backend module will follow.

Resolves: #104371
Releases: main
Change-Id: I030e12cb144befb267dea1c3d5770787c8480c35
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85231
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>