[TASK] Properly set bcrypt cost
BcryptPasswordHash has an option to set a cost
value. The lower barrier used to be identical
to PHP constant PASSWORD_BCRYPT_DEFAULT_COST.

Tests use the lower barrier to speed up tests.
With PHP <8.4, PASSWORD_BCRYPT_DEFAULT_COST is
10, it has been raised to 12 with PHP 8.4.

The patch frees our code from access to this
constant and hard codes 10 as lower barrier
check. This keeps the API as is, and this hash
is by default set to 12 already in default
options if none are provided.

Resolves: #104373
Releases: main, 12.4, 11.5
Change-Id: I4c1e94fdc50aa3075570549ccdd3b81a2de2e98a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85235
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>