[BUGFIX] Pass request in ext:form preview mode
The ArrayFormFactory and underlying functionality
requires a PSR-7 Request object. In backend context,
when rendering the form element preview, the
request is now properly passed.

Additionally, the `RenderContentElementPreviewViewHelper`,
used in the preview mode to render the content
element preview initializes the PageLayoutContext,
which since #103326 also requires a PSR-7 Request.

Both places are adjusted accordingly, resolving the
TypeErrors and making the form preview working again.

Resolves: #104375
Releases: main
Change-Id: Ib0e07b8a38cca1bb0d78f72b35525e36814e2521
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85242
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Elias Häußler <e.haeussler@familie-redlich.de>