[BUGFIX] Prevent broken page module state
In case content elements exist in the database,
having a CType for which no TCA exists, e.g.
because an extension providing this CType
has been deactivated, no more UndefinedSchemaException,
is thrown. This allows an editor to properly
handle the situation as before.

Note: Such elements are displayed with an
"INVALID VALUE ("INVALID_CTYPE")" badge.

Resolves: #104448
Releases: main
Change-Id: I94cd9c42040274536730b020ae29bf9ff13bb3f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85322
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>