[BUGFIX] Free ext:form ConfirmationFinisher COR calls from extbase
* ext:form is FE an extbase plugin and operates with an
  extbase request created by extbase bootstrap (or by
  FluidTemplateContentObject, but that's a different story).
* ConfirmationFinisher has option 'contentElementUid', which
  - when set - creates a ContentObjectRenderer to render
  specific content elements.
* ContentObjectRenderer then often renders this content
  element using FluidTemplateContentObject. Various code
  within Fluid rendering then applies magic to create new
  requests, *if* an extbase request is given. This can lead
  to broken template resolving when rendering in
  FluidTemplateContentObject.

This entire thing is quite messy. We're taking measures to
resolve the underlying flaws in v13. To fix the situation
in v12, the idea is like that: When ConfirmationFinisher
calls ContentObjectRenderer and hands over its request, it
should "free" that request from its own extbase context
to not trigger pollution by code that does things if an
extbase request is given.

Note this was not an issue in v11 since only v12 started
handing over the extbase request to ContentObjectRenderer,
in v11 it was indirectly pulled from $GLOBALS['TYPO3_REQUEST']
within ContentObjectRenderer, which is the "original" request
before it is turned into an extbase request by extbase
plugin bootstrap.

Resolves: #104462
Related: #100623
Related: #104472
Related: #104471
Related: #104473
Releases: main, 12.4
Change-Id: Id53bc9f159dc2575a8bedb63cd7e7a039864f92b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85336
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Simon Praetorius <simon@praetorius.me>