[TASK] Use recursive common table expressions in RootlineUtility
The patch implements a first use of a recursive
common table expression ("CTE" / "WITH RECURSIVE").

The target of this first CTE is RootlineUtility.
RootlineUtility tends to execute many single queries,
especially in the Frontend with not fully cached
page requests.

The job of RootlineUtility->get() is to find the path
of a page up to its tree root.

While a basic CTE to solve this recursively with a
single query is relatively straight forward, the
true magic unfolds when workspace overlays are
taken into account: The final query is very
efficient, but far from trivial and nothing for the
faint of heart. Note this CTE is capable of resolving
workspace scenarios with one query, but language overlay
scenarios are not yet implemented.

Core v13 already improved RootlineUtility test coverage
with previous patches, and a lot of functional patches
indirectly test rootline determination as well. All in
all, we are confident this CTE solution works well.

With given solution, the number of queries already
shrinks quite significantly, currently up to 1/4.
Note this patch is the first one in this direction:
There are further possible optimization steps, patches
dealing with more details will follow.

Resolves: #104506
Related: #104013
Related: #104491
Releases: main
Change-Id: Ib82516ebe9c5198c5ead26d82f6b4e60d11e7b95
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85374
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>