[BUGFIX] Fix variable key loop to avoid DB compare issues
With #93223 minor index handling within the Database
Analyzer Stack was fixed.

Via DBAL 4 changes in "main" (and not 12.4),
some variable names in loops were changed. When
backporting #93223 to 12.4, these changes were
not carried over, which led to an inner loop
re-using a variable name from the outer loop.

Due to this, the index name generation could not
properly recognize all renamed keys, and only
the last renamed key was recognized in the
Database Analyzer.

This patch adjusts the inner and outer loop
variable names to be unique and in line with
what is in place in "main".

Resolves: #104449
Related: #93223
Releases: 12.4
Change-Id: Id2aeefad661188c1fe458c773f571abdf9bfb0bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85379
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Sergio Catalá <cajondesartre@gmail.com>
Reviewed-by: Sergio Catalá <cajondesartre@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>