[BUGFIX] Skip loading fallback icon
Since #102834 content elements and plugins are auto registered for the
"New Content Element Wizard" but if no icon is provided, a broken icon
has been shown. To avoid a broken state, the fallback icon is removed.

It is better to show no icon instead of guessing and failing with a
fallback icon.

Related: #102834
Resolves: #103663
Releases: main
Change-Id: I14b49a3f956d15aea37aa2e57b8062076b04a6ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85396
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>