[TASK] Use country API in site_language's flag field
Instead of hard-coding all countries, the country API can be used.
The main benefit is now that the name instead of ISO2 of the country is
shown, which is especially useful when dealing with not well-known
countries.

Furthermore, the item groups are used to differ between real countries,
colors and some miscellaneous items.

Resolves: #104517
Releases: main
Change-Id: I411c29ffec8dba2b07ed91327681860d29f859ae
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85399
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>